move getTierName to separate file
